total now returns to 2 decimal places

+++ app/src/main/java/com/example/user/bills/Item.java
@@ -1,14 +1,16 @@
 package com.example.user.bills;
 
+import java.math.BigDecimal;
+
 /**
  * Created by user on 23/08/2016.
  */
 public class Item {
 
     private String description;
-    private int cost;
+    private float cost;
 
-    public Item(String description, int cost) {
+    public Item(String description, float cost) {
         this.description = description;
         this.cost = cost;
     }
@@ -21,7 +23,7 @@ public class Item {
         this.description = newDescription;
     }
 
-    public int getItemCost() {
+    public float getItemCost() {
         return this.cost;
     }
